% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTShiny2.R
\name{CTShiny2}
\alias{CTShiny2}
\title{Launch 'CTShiny2' Interface}
\usage{
CTShiny2()
}
\value{
Nothing
}
\description{
CTShiny2() loads interactive user interface built using R 'shiny'.
}
\details{
Start CTShiny2

The interactive user interface is to provide an easy way for classification tree analysis and downloading relevant plot.
}
\examples{
if(interactive()){
library(rmarkdown)
CTShiny2()
}
}
\keyword{CTShiny2}
